[elk] Refactor sortinghat module name
To avoid conflicts with sortinghat module within sortinghat
package.
