+++ grimoire_elk/elk/sortinghat_gelk.py
