+++ utils/gelk.py
@@ -34,7 +34,7 @@ from grimoire_elk.elk.elastic import ElasticConnectException
 from grimoire_elk.elk.bugzilla import BugzillaEnrich
 from grimoire_elk.elk.gerrit import GerritEnrich
 from grimoire_elk.elk.github import GitHubEnrich
-from grimoire_elk.elk.sortinghat import SortingHat
+from grimoire_elk.elk.sortinghat_gelk import SortingHat
 
 from grimoire_elk.ocean.bugzilla import BugzillaOcean
 from grimoire_elk.ocean.gerrit import GerritOcean
